requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
amqp==5.1.1
APScheduler==3.9.1
asgiref==3.4.1
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.1.2
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
cryptography==3.4.8
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-celery-beat==2.2.1
django-crontab==0.7.1
django-timezone-field==4.2.3
django4-background-tasks==1.2.7
djangorestframework==3.13.1
eight==1.0.1
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
Jinja2==3.1.1
kombu==5.2.4
lxml==4.8.0
MarkupSafe==2.1.1
multidict==6.0.2
openleadr==0.5.24
packaging==21.3
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
pyOpenSSL==19.1.0
pyparsing==3.0.8
python-crontab==2.3.4
python-dateutil==2.8.2
python-dotenv==0.19.2
pytz==2021.3
pytz-deprecation-shim==0.1.0.post0
redis==4.2.2
requests==2.27.1
signxml-openadr==2.9.1
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
tzlocal==4.2
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
whitenoise==5.3.0
wrapt==1.14.0
xmltodict==0.12.0
yarl==1.7.2